fix: authorization issue with google oauth This issue only appears for users who already have an account registered in the database
Google provides only a 1 refresh token when the user signs in for the first time. This forces the user to remove the app from their google permissions settings to be able to login again.

After consulting the Auth.js documentation I was able to fix the issue.

for more information: https://authjs.dev/getting-started/providers/google#refresh-token